Refactor Admin Bundle event classes (#14744)
* Refactor Admin Bundle event classes - related to #14642

* Refactor Admin Bundle event classes - related to #14642

* Refactor Admin Bundle event classes - related to #14642

* Refactor Admin Bundle event classes - related to #14642

* upgrade notes

* fix AdminEvents

* Add missing old classes

* Docs
