+++ bundles/AdminBundle/Event/IndexActionSettingsEvent.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Pimcore
+ *
+ * This source file is available under two different licenses:
+ * - GNU General Public License version 3 (GPLv3)
+ * - Pimcore Commercial License (PCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ *  @license    http://www.pimcore.org/license     GPLv3 and PCL
+ */
+
+namespace Pimcore\Bundle\AdminBundle\Event;
+
+use Symfony\Contracts\EventDispatcher\Event;
+
+class IndexActionSettingsEvent extends Event
+{
+    /**
+     * @var array
+     */
+    private $settings;
+
+    public function __construct(array $settings)
+    {
+        $this->settings = $settings;
+    }
+
+    public function getSettings(): array
+    {
+        return $this->settings;
+    }
+
+    public function setSettings(array $settings)
+    {
+        $this->settings = $settings;
+    }
+
+    /**
+     * @param string $key
+     * @param mixed $value
+     */
+    public function addSetting(string $key, $value)
+    {
+        $this->settings[$key] = $value;
+    }
+}
