+++ bundles/AdminBundle/Event/Login/LoginCredentialsEvent.php
@@ -1,0 +1,56 @@
+<?php
+
+/**
+ * Pimcore
+ *
+ * This source file is available under two different licenses:
+ * - GNU General Public License version 3 (GPLv3)
+ * - Pimcore Commercial License (PCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ *  @license    http://www.pimcore.org/license     GPLv3 and PCL
+ */
+
+namespace Pimcore\Bundle\AdminBundle\Event\Login;
+
+use Pimcore\Event\Traits\RequestAwareTrait;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Contracts\EventDispatcher\Event;
+
+class LoginCredentialsEvent extends Event
+{
+    use RequestAwareTrait;
+
+    /**
+     * @var array
+     */
+    protected $credentials;
+
+    /**
+     * @param Request $request
+     * @param array $credentials
+     */
+    public function __construct(Request $request, array $credentials)
+    {
+        $this->request = $request;
+        $this->credentials = $credentials;
+    }
+
+    /**
+     * @return array
+     */
+    public function getCredentials()
+    {
+        return $this->credentials;
+    }
+
+    /**
+     * @param array $credentials
+     */
+    public function setCredentials(array $credentials)
+    {
+        $this->credentials = $credentials;
+    }
+}
