+++ bundles/AdminBundle/Event/Login/LoginRedirectEvent.php
@@ -1,0 +1,73 @@
+<?php
+
+/**
+ * Pimcore
+ *
+ * This source file is available under two different licenses:
+ * - GNU General Public License version 3 (GPLv3)
+ * - Pimcore Commercial License (PCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ *  @license    http://www.pimcore.org/license     GPLv3 and PCL
+ */
+
+namespace Pimcore\Bundle\AdminBundle\Event\Login;
+
+use Symfony\Contracts\EventDispatcher\Event;
+
+class LoginRedirectEvent extends Event
+{
+    /**
+     * @var string
+     */
+    protected $routeName;
+
+    /**
+     * @var array
+     */
+    protected $routeParams;
+
+    /**
+     * @param string $routeName
+     * @param array $routeParams
+     */
+    public function __construct(string $routeName, array $routeParams = [])
+    {
+        $this->routeName = $routeName;
+        $this->routeParams = $routeParams;
+    }
+
+    /**
+     * @return string
+     */
+    public function getRouteName(): string
+    {
+        return $this->routeName;
+    }
+
+    /**
+     * @param string $routeName
+     */
+    public function setRouteName(string $routeName): void
+    {
+        $this->routeName = $routeName;
+    }
+
+    /**
+     * @return array
+     */
+    public function getRouteParams(): array
+    {
+        return $this->routeParams;
+    }
+
+    /**
+     * @param array $routeParams
+     */
+    public function setRouteParams(array $routeParams): void
+    {
+        $this->routeParams = $routeParams;
+    }
+}
