+++ bundles/AdminBundle/Event/Login/LogoutEvent.php
@@ -1,0 +1,51 @@
+<?php
+
+/**
+ * Pimcore
+ *
+ * This source file is available under two different licenses:
+ * - GNU General Public License version 3 (GPLv3)
+ * - Pimcore Commercial License (PCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ *  @license    http://www.pimcore.org/license     GPLv3 and PCL
+ */
+
+namespace Pimcore\Bundle\AdminBundle\Event\Login;
+
+use Pimcore\Event\Traits\RequestAwareTrait;
+use Pimcore\Event\Traits\ResponseAwareTrait;
+use Pimcore\Model\User;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Contracts\EventDispatcher\Event;
+
+class LogoutEvent extends Event
+{
+    use RequestAwareTrait;
+    use ResponseAwareTrait;
+
+    /**
+     * @var User
+     */
+    protected $user;
+
+    /**
+     * @param Request $request
+     * @param User $user
+     */
+    public function __construct(Request $request, User $user)
+    {
+        $this->request = $request;
+        $this->user = $user;
+    }
+
+    /**
+     * @return User
+     */
+    public function getUser()
+    {
+        return $this->user;
+    }
+}
