+++ bundles/AdminBundle/Event/Model/AssetDeleteInfoEvent.php
@@ -13,15 +13,17 @@
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Event\Model;
+namespace Pimcore\Bundle\AdminBundle\Event\Model;
+
+use Pimcore\Bundle\AdminBundle\Event\Traits\ElementDeleteInfoEventTrait;
+use Pimcore\Event\Model\AssetEvent;
 
 /**
- * Class DataObjectDeleteInfoEvent
+ * Class AssetDeleteInfoEvent
  *
  * @package Pimcore\Event\Model
- *
- * @deprecated and will be removed in Pimcore 11. Use Pimcore\Bundle\AdminBundle\Event\Model\DataObjectDeleteInfoEvent instead
  */
-class DataObjectDeleteInfoEvent extends \Pimcore\Bundle\AdminBundle\Event\Model\DataObjectDeleteInfoEvent
+class AssetDeleteInfoEvent extends AssetEvent implements ElementDeleteInfoEventInterface
 {
+    use ElementDeleteInfoEventTrait;
 }
