+++ doc/Development_Documentation/18_Tools_and_Features/02_Custom_Icons.md
@@ -13,7 +13,7 @@ be changed.
  
 The basic idea is to provide one's own implementation of `Pimcore\Model\Element\AdminStyle`.
  
-This can be achieved by attaching a listener to the [`AdminEvents::RESOLVE_ELEMENT_ADMIN_STYLE`](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/AdminEvents.php#L396-L407) event. 
+This can be achieved by attaching a listener to the [`AdminEvents::RESOLVE_ELEMENT_ADMIN_STYLE`](https://github.com/pimcore/pimcore/blob/10.5/bundles/AdminBundle/Event/AdminEvents.php#L396-L407) event. 
 
 Example:
