+++ doc/Development_Documentation/20_Extending_Pimcore/11_Event_API_and_Event_Manager.md
@@ -33,7 +33,7 @@ All Pimcore events are defined and documented as a constant on component specifi
 - [Mail](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/MailEvents.php)
 - [Notifications](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/NotificationEvents.php)
 - [Redirect](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/RedirectEvents.php)
-- [Admin](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/AdminEvents.php)
+- [Admin](https://github.com/pimcore/pimcore/blob/10.5/bundles/AdminBundle/Event/AdminEvents.php)
 - [Frontend](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/FrontendEvents.php)
 - [Cache](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/CoreCacheEvents.php)
 - [Full-Page Cache](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/FullPageCacheEvents.php)
@@ -108,7 +108,7 @@ To ensure maximum security, it is advisable to combine this with an object DI to
 ### Hook into the Open Document|Asset|Data Object dialog
 
 By the default, Pimcore tries to a resolve an element by its ID or path.
-You can change this behavior by handling the [AdminEvents::RESOLVE_ELEMENT](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/AdminEvents.php) event
+You can change this behavior by handling the [AdminEvents::RESOLVE_ELEMENT](https://github.com/pimcore/pimcore/blob/10.5/bundles/AdminBundle/Event/AdminEvents.php) event
 and implement your own logic.
 
 ```php
