+++ doc/Development_Documentation/26_Best_Practice/60_Modifying_Permissions_based_on_Object_Data.md
@@ -1,6 +1,6 @@
 # Modifying Permissions based on Object Data
 
-The event [`OBJECT_GET_PRE_SEND_DATA`](https://github.com/pimcore/pimcore/blob/10.5/lib/Event/AdminEvents.php#L292-L304)
+The event [`OBJECT_GET_PRE_SEND_DATA`](https://github.com/pimcore/pimcore/blob/10.5/bundles/AdminBundle/Event/AdminEvents.php#L292-L304)
 can be used to manipulate the server response before object data is sent to Pimcore Backend UI when opening the detail
 view of an Pimcore object.
