+++ lib/Event/Admin/Login/LogoutEvent.php
@@ -15,37 +15,9 @@
 
 namespace Pimcore\Event\Admin\Login;
 
-use Pimcore\Event\Traits\RequestAwareTrait;
-use Pimcore\Event\Traits\ResponseAwareTrait;
-use Pimcore\Model\User;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Contracts\EventDispatcher\Event;
-
-class LogoutEvent extends Event
+/**
+ * @deprecated and will be removed in Pimcore 11. Use Pimcore\Bundle\AdminBundle\Event\Login\LogoutEvent instead
+ */
+class LogoutEvent extends \Pimcore\Bundle\AdminBundle\Event\Login\LogoutEvent
 {
-    use RequestAwareTrait;
-    use ResponseAwareTrait;
-
-    /**
-     * @var User
-     */
-    protected $user;
-
-    /**
-     * @param Request $request
-     * @param User $user
-     */
-    public function __construct(Request $request, User $user)
-    {
-        $this->request = $request;
-        $this->user = $user;
-    }
-
-    /**
-     * @return User
-     */
-    public function getUser()
-    {
-        return $this->user;
-    }
 }
